Neet AI/requirements.txt
python-dotenv
streamlit
Pillow
crewai
langchain
langchain-community
faiss-cpu
sentence-transformers
google-generativeai
python-pdf